+++ spec/controllers/api/v2/account_lists/pledges_controller_spec.rb
@@ -65,6 +65,23 @@ RSpec.describe Api::V2::AccountLists::PledgesController, type: :controller do
   # spec/support/shared_controller_examples/*
   include_examples 'index_examples'
 
+  describe 'filtering' do
+    before { api_login(user) }
+
+    context 'status filter' do
+      let(:contact2) { create(:contact, account_list: account_list) }
+      let!(:pledge3) { create(:pledge, account_list: account_list, contact: contact2, amount: 10.00, status: :processed) }
+
+      it 'filters results' do
+        get :index, parent_param.merge(filter: { status: 'processed' })
+
+        expect(response.status).to eq(200)
+        expect(response_json['data'].length).to eq(1)
+        expect(response_json['data'][0]['id']).to eq(pledge3.uuid)
+      end
+    end
+  end
+
   include_examples 'show_examples'
 
   include_examples 'create_examples'